Move probe related functions to ftsprobe.c
Also break the processing of response and updating of cluster configuration
into two separate functions as suggested by Ashwin.
