+++ src/backend/fts/fts.c
@@ -170,11 +170,6 @@ ftsprobe_forkexec(void)
 }
 #endif   /* EXEC_BACKEND */
 
-inline bool
-IsFtsShudownRequested(void) {
-	return shutdown_requested;
-}
-
 static void
 RequestShutdown(SIGNAL_ARGS)
 {
@@ -387,11 +382,11 @@ CdbComponentDatabases *readCdbComponentInfoAndUpdateStatus(MemoryContext probeCo
 	return cdbs;
 }
 
-static void
+void
 probeWalRepUpdateConfig(int16 dbid, int16 segindex, char role,
 						bool IsSegmentAlive, bool IsInSync)
 {
-	Assert(IsInSync ? IsSegmentAlive : true);
+	AssertImply(IsInSync, IsSegmentAlive);
 
 	/*
 	 * Insert new tuple into gp_configuration_history catalog.
@@ -485,369 +480,6 @@ probeWalRepUpdateConfig(int16 dbid, int16 segindex, char role,
 		pfree(newtuple);
 
 		heap_close(configrel, RowExclusiveLock);
-	}
-}
-
-/*
- * Timeout is said to have occurred if greater than gp_fts_probe_timeout
- * seconds have elapsed since connection start and a response is not received.
- * Segments for which a response is received already or that have failed with
- * all retries exhausted are exempted from timeout evaluation.
- */
-static void
-ftsCheckTimeout(per_segment_info *ftsInfo, pg_time_t now)
-{
-	if (!IsFtsProbeStateSuccess(ftsInfo->state) &&
-		ftsInfo->retry_count < gp_fts_probe_retries &&
-		(int) (now - ftsInfo->startTime) > gp_fts_probe_timeout)
-	{
-		elog(LOG,
-			 "FTS timeout detected for (content=%d, dbid=%d) "
-			 "state=%d, retry_count=%d,",
-			 ftsInfo->primary_cdbinfo->segindex,
-			 ftsInfo->primary_cdbinfo->dbid, ftsInfo->state,
-			 ftsInfo->retry_count);
-		ftsInfo->state = nextFailedState(ftsInfo->state);
-		/* For now no more retries if connection timed out. */
-
-		ftsInfo->retry_count = gp_fts_probe_retries;
-	}
-}
-
-/*
- * Return true for segments whose response is ready to be processed.  Segments
- * whose response is already processed should have response->conn set to NULL.
- */
-static bool
-ftsResponseReady(per_segment_info *ftsInfo)
-{
-	return ((IsFtsProbeStateSuccess(ftsInfo->state) ||
-			 IsFtsProbeStateFailed(ftsInfo->state)) && ftsInfo->conn != NULL);
-}
-
-/*
- * Process resonses from primary segments:
- * (a) Transition internal state so that segments can be messaged subsequently
- * (e.g. promotion and turning off syncrep).
- * (b) Update gp_segment_configuration catalog table, if needed.
- */
-bool
-probeWalRepPublishUpdate(fts_context *context)
-{
-	bool is_updated = false;
-
-	pg_time_t now = (pg_time_t) time(NULL);
-
-	for (int response_index = 0; response_index < context->num_pairs;
-		 response_index ++)
-	{
-		per_segment_info *ftsInfo = &(context->perSegInfos[response_index]);
-
-		ftsCheckTimeout(ftsInfo, now);
-		/*
-		 * Consider segments that are in final state (success / failure) and
-		 * that are not already processed.
-		 */
-		if (!ftsResponseReady(ftsInfo))
-			continue;
-		/*
-		 * We must be in a final state, which is (1) success or (2) failure
-		 * with retries exhausted.
-		 */
-		Assert(IsFtsProbeStateSuccess(ftsInfo->state) ||
-			   IsFtsProbeStateFailed(ftsInfo->state));
-		AssertImply(IsFtsProbeStateFailed(ftsInfo->state),
-					ftsInfo->retry_count == gp_fts_probe_retries);
-
-		/*
-		 * Before sending promote message, primary_cdbinfo refence is swapped
-		 * with the mirror, leaving primary_cdbinfo->role to be 'm'.  So the
-		 * assertion applies only to non-promote cases.
-		 */
-		AssertImply(ftsInfo->state != FTS_PROMOTE_SUCCESS &&
-					ftsInfo->state != FTS_PROMOTE_FAILED,
-					SEGMENT_IS_ACTIVE_PRIMARY(ftsInfo->primary_cdbinfo));
-
-		if (!FtsIsActive())
-			return false;
-
-		CdbComponentDatabaseInfo *primary = ftsInfo->primary_cdbinfo;
-
-		CdbComponentDatabaseInfo *mirror = ftsInfo->mirror_cdbinfo;
-
-		bool UpdatePrimary = false;
-		bool UpdateMirror = false;
-
-		/*
-		 * Currently, mode of primary and mirror should be same, either both
-		 * reflecting in sync or not in sync.
-		 */
-		Assert(primary->mode == mirror->mode);
-
-		bool IsPrimaryAlive = ftsInfo->result.isPrimaryAlive;
-		/* Trust a response from primary only if it's alive. */
-		bool IsMirrorAlive =  IsPrimaryAlive ?
-			ftsInfo->result.isMirrorAlive : SEGMENT_IS_ALIVE(mirror);
-		bool IsInSync = IsPrimaryAlive ?
-			ftsInfo->result.isInSync : false;
-
-		/* If primary and mirror are in sync, then both have to be ALIVE. */
-		AssertImply(IsInSync, IsPrimaryAlive && IsMirrorAlive);
-		/* Primary must enable syncrep as long as it thinks mirror is alive. */
-		AssertImply(IsMirrorAlive && IsPrimaryAlive,
-					ftsInfo->result.isSyncRepEnabled);
-
-		/* Only swapped in promotion; by default, keep the current roles. */
-		char newPrimaryRole = GP_SEGMENT_CONFIGURATION_ROLE_PRIMARY;
-		char newMirrorRole = GP_SEGMENT_CONFIGURATION_ROLE_MIRROR;
-
-		/*
-		 * Allow updates to configuration upon receiving response to PROBE
-		 * message.  Any other message (SYNCREP_OFF / PROMOTE) should not
-		 * result in configuration update.
-		 *
-		 * XXX For SYNCREP_OFF message, a primary may respond with mirror is
-		 * back up and syncrep is turned on.  In that case, the mirror should
-		 * be marked back up in coniguration.
-		 */
-		if (ftsInfo->state == FTS_PROBE_SUCCESS ||
-			ftsInfo->state == FTS_PROBE_FAILED)
-		{
-			UpdatePrimary = (IsPrimaryAlive != SEGMENT_IS_ALIVE(primary));
-			UpdateMirror = (IsMirrorAlive != SEGMENT_IS_ALIVE(mirror));
-			/*
-			 * If probe response state is different from current state in
-			 * configuration, update both primary and mirror.
-			 */
-			if (IsInSync != SEGMENT_IS_IN_SYNC(primary))
-				UpdatePrimary = UpdateMirror = true;
-		}
-
-		switch(ftsInfo->state)
-		{
-			case FTS_PROBE_SUCCESS:
-				Assert(IsPrimaryAlive);
-				if (ftsInfo->result.isSyncRepEnabled && !IsMirrorAlive)
-				{
-					/*
-					 * Primaries that have syncrep enabled continue to block
-					 * commits.  FTS must notify them to unblock commits by
-					 * sending syncrep off message.
-					 */
-					ftsInfo->state = FTS_SYNCREP_SEGMENT;
-					elog(LOG, "FTS turning syncrep off on (content=%d, dbid=%d)",
-						 primary->segindex, primary->dbid);
-					/*
-					 * Mirror must be marked down in FTS configuration before
-					 * primary can be notified to unblock commits.
-					 */
-					Assert(UpdateMirror || !SEGMENT_IS_ALIVE(mirror));
-				}
-				else if (ftsInfo->result.isRoleMirror && !IsMirrorAlive)
-				{
-					/*
-					 * A promote message sent previously didn't make it to the
-					 * mirror.  Catalog must have been updated before sending
-					 * the previous promote message.
-					 */
-					Assert(!SEGMENT_IS_ALIVE(mirror));
-					Assert(SEGMENT_IS_NOT_INSYNC(mirror));
-					Assert(SEGMENT_IS_NOT_INSYNC(primary));
-					Assert(!ftsInfo->result.isSyncRepEnabled);
-					Assert(!UpdateMirror);
-					Assert(!UpdatePrimary);
-					elog(LOG, "FTS resending promote request to (content=%d,"
-						 " dbid=%d)", primary->segindex, primary->dbid);
-					ftsInfo->state = FTS_PROMOTE_SEGMENT;
-				}
-				break;
-			case FTS_PROBE_FAILED:
-				/* Primary is down, see if mirror can be promoted. */
-				Assert(!IsPrimaryAlive);
-				if (SEGMENT_IS_IN_SYNC(mirror))
-				{
-					/*
-					 * Primary must have been recorded as in-sync before the
-					 * probe.
-					 */
-					Assert(SEGMENT_IS_IN_SYNC(primary));
-					/*
-					 * Swap the primary and mirror references so that the
-					 * mirror will be promoted in subsequent connect, poll,
-					 * send, receive steps.
-					 */
-					elog(LOG, "FTS promoting mirror (content=%d, dbid=%d) to be"
-						 " the new primary", mirror->segindex, mirror->dbid);
-					ftsInfo->state = FTS_PROMOTE_SEGMENT;
-					ftsInfo->primary_cdbinfo = mirror;
-					ftsInfo->mirror_cdbinfo = primary;
-
-					/*
-					 * Flip the roles and mark the failed primary as down in
-					 * FTS configuration before sending promote message.
-					 * Dispatcher should no longer consider the failed primary
-					 * for gang creation, FTS should no longer probe the failed
-					 * primary.
-					 */
-					newPrimaryRole = GP_SEGMENT_CONFIGURATION_ROLE_MIRROR;
-					newMirrorRole = GP_SEGMENT_CONFIGURATION_ROLE_PRIMARY;
-				}
-				else
-				{
-					elog(WARNING, "FTS double fault detected (content=%d) "
-						 "primary dbid=%d, mirror dbid=%d",
-						 primary->segindex, primary->dbid, mirror->dbid);
-					UpdatePrimary = UpdateMirror = false;
-				}
-				break;
-			case FTS_SYNCREP_FAILED:
-				/*
-				 * XXX transactions may block forever, waiting for a mirror
-				 * that doesn't exist.  Should we PANIC here?
-				 */
-				elog(WARNING, "FTS failed to turn off syncrep on (content=%d,"
-					 " dbid=%d)", primary->segindex, primary->dbid);
-				break;
-			case FTS_PROMOTE_FAILED:
-				elog(WARNING, "FTS double fault detected (content=%d) "
-					 "primary dbid=%d, mirror dbid=%d",
-					 primary->segindex, primary->dbid, mirror->dbid);
-				break;
-			case FTS_PROMOTE_SUCCESS:
-				Assert(ftsInfo->result.isRoleMirror);
-				elog(LOG, "FTS mirror (content=%d, dbid=%d) promotion "
-					 "triggerred successfully",
-					 primary->segindex, primary->dbid);
-				break;
-			case FTS_SYNCREP_SUCCESS:
-				elog(LOG, "FTS primary (content=%d, dbid=%d) notified to turn "
-					 "syncrep off", primary->segindex, primary->dbid);
-				break;
-			default:
-				elog(ERROR, "FTS invalid internal state %d for (content=%d)"
-					 "primary dbid=%d, mirror dbid=%d", ftsInfo->state,
-					 primary->segindex, primary->dbid, mirror->dbid);
-				break;
-		}
-		/* Close connection and reset result for next message, if any. */
-		memset(&ftsInfo->result, 0, sizeof(probe_result));
-		PQfinish(ftsInfo->conn);
-		ftsInfo->conn = NULL;
-		ftsInfo->poll_events = ftsInfo->poll_revents = 0;
-		ftsInfo->retry_count = 0;
-
-		if (shutdown_requested)
-		{
-			elog(LOG, "shutdown in progress, ignoring FTS prober updates");
-			return is_updated;
-		}
-
-		if (UpdatePrimary || UpdateMirror)
-		{
-			/*
-			 * Commit/abort transaction below will destroy
-			 * CurrentResourceOwner.  We need it for catalog reads.
-			 */
-			ResourceOwner save = CurrentResourceOwner;
-			StartTransactionCommand();
-			GetTransactionSnapshot();
-
-			if (UpdatePrimary)
-				probeWalRepUpdateConfig(primary->dbid, primary->segindex,
-										newPrimaryRole, IsPrimaryAlive,
-										IsInSync);
-
-			if (UpdateMirror)
-				probeWalRepUpdateConfig(mirror->dbid, mirror->segindex,
-										newMirrorRole, IsMirrorAlive,
-										IsInSync);
-
-			is_updated = true;
-
-			CommitTransactionCommand();
-			CurrentResourceOwner = save;
-
-			/*
-			 * Update the status to in-memory variable as well used by
-			 * dispatcher, now that changes has been persisted to catalog.
-			 */
-			Assert(ftsProbeInfo);
-			/* XXX why shouldn't FtsControlBlock->ControlLock be held here? */
-			if (IsPrimaryAlive)
-				FTS_STATUS_SET_UP(ftsProbeInfo->fts_status[primary->dbid]);
-			else
-				FTS_STATUS_SET_DOWN(ftsProbeInfo->fts_status[primary->dbid]);
-
-			if (IsMirrorAlive)
-				FTS_STATUS_SET_UP(ftsProbeInfo->fts_status[mirror->dbid]);
-			else
-				FTS_STATUS_SET_DOWN(ftsProbeInfo->fts_status[mirror->dbid]);
-		}
-	}
-
-	return is_updated;
-}
-
-static void
-FtsWalRepInitProbeContext(CdbComponentDatabases *cdbs, fts_context *context)
-{
-	context->num_pairs = cdbs->total_segments;
-	context->perSegInfos = (per_segment_info *) palloc0(
-		context->num_pairs * sizeof(per_segment_info));
-
-	int response_index = 0;
-
-	for(int segment_index = 0; segment_index < cdbs->total_segment_dbs; segment_index++)
-	{
-		CdbComponentDatabaseInfo *segment = &(cdbs->segment_db_info[segment_index]);
-		per_segment_info *ftsInfo = &(context->perSegInfos[response_index]);
-
-		if (!SEGMENT_IS_ACTIVE_PRIMARY(segment))
-			continue;
-
-		/*
-		 * We initialize the probe_result.IsPrimaryAlive and
-		 * probe_result.IsMirrorAlive to current state, and that will prevent
-		 * changing of mirror status if the primary goes down.
-		 */
-		CdbComponentDatabaseInfo *primary = segment;
-		CdbComponentDatabaseInfo *mirror = FtsGetPeerSegment(cdbs,
-															 primary->segindex,
-															 primary->dbid);
-
-		/*
-		 * If there is no mirror under this primary, no need to probe.
-		 */
-		if (!mirror)
-		{
-			context->num_pairs--;
-			continue;
-		}
-
-		/* primary in catalog will NEVER be marked down. */
-		Assert(FtsIsSegmentAlive(primary));
-
-		/*
-		 * Before we probe, we always DEFAULT the response for primary is
-		 * down. This will only be changed to true if primary is alive,
-		 * otherwise some error happened during the probe. Similarly, always
-		 * DEFAULT the response to not in sync. Only if primary communicates
-		 * positively, consider it in sync.
-		 */
-		ftsInfo->result.isPrimaryAlive = false;
-		ftsInfo->result.isMirrorAlive = false;
-		ftsInfo->result.isInSync = false;
-		ftsInfo->result.isSyncRepEnabled = false;
-		ftsInfo->result.retryRequested = false;
-		ftsInfo->result.isRoleMirror = false;
-		ftsInfo->result.dbid = primary->dbid;
-		ftsInfo->state = FTS_PROBE_SEGMENT;
-
-		ftsInfo->primary_cdbinfo = primary;
-		ftsInfo->mirror_cdbinfo = mirror;
-
-		Assert(response_index < context->num_pairs);
-		response_index ++;
 	}
 }
 
@@ -947,13 +579,8 @@ void FtsLoop()
 		 * stuff palloc()s internally
 		 */
 		oldContext = MemoryContextSwitchTo(probeContext);
-		initPollFds(cdbs->total_segments);
 
-		fts_context context;
-
-		FtsWalRepInitProbeContext(cdbs, &context);
-
-		updated_probe_state = FtsWalRepMessageSegments(&context);
+		updated_probe_state = FtsWalRepMessageSegments(cdbs);
 
 		MemoryContextSwitchTo(oldContext);
 
@@ -1002,38 +629,4 @@ bool
 FtsIsActive(void)
 {
 	return (!skipFtsProbe && !shutdown_requested);
-}
-
-bool
-FtsIsSegmentAlive(CdbComponentDatabaseInfo *segInfo)
-{
-	if (SEGMENT_IS_ACTIVE_MIRROR(segInfo) && SEGMENT_IS_ALIVE(segInfo))
-		return true;
-
-	if (SEGMENT_IS_ACTIVE_PRIMARY(segInfo))
-		return true;
-
-	return false;
-}
-
-/*
- * Get peer segment descriptor
- */
-CdbComponentDatabaseInfo *FtsGetPeerSegment(CdbComponentDatabases *cdbs,
-											int content, int dbid)
-{
-	int i;
-
-	for (i=0; i < cdbs->total_segment_dbs; i++)
-	{
-		CdbComponentDatabaseInfo *segInfo = &cdbs->segment_db_info[i];
-
-		if (segInfo->segindex == content && segInfo->dbid != dbid)
-		{
-			/* found it */
-			return segInfo;
-		}
-	}
-
-	return NULL;
 }
