+++ src/backend/fts/test/Makefile
@@ -2,7 +2,7 @@ subdir=src/backend/fts
 top_builddir=../../../..
 include $(top_builddir)/src/Makefile.global
 
-TARGETS=ftsmessagehandler ftsprobe fts
+TARGETS=ftsmessagehandler ftsprobe
 
 include $(top_builddir)/src/backend/mock.mk
 
@@ -17,20 +17,9 @@ ftsmessagehandler.t: \
     $(MOCK_DIR)/backend/postmaster/postmaster_mock.o \
     $(MOCK_DIR)/backend/access/transam/xlog_mock.o
 
-ftsprobe.t: fts_test_helper.o \
-    $(MOCK_DIR)/backend/libpq/fe-exec_mock.o \
-    $(MOCK_DIR)/backend/libpq/fe-connect_mock.o \
-    $(MOCK_DIR)/backend/cdb/cdbthreadlog_mock.o
-
-
-fts.t: fts_test_helper.o \
+ftsprobe.t: \
+    $(MOCK_DIR)/backend/fts/fts_mock.o \
     $(MOCK_DIR)/backend/libpq/fe-exec_mock.o \
     $(MOCK_DIR)/backend/libpq/fe-connect_mock.o \
     $(MOCK_DIR)/backend/access/transam/xact_mock.o \
-    $(MOCK_DIR)/backend/utils/misc/faultinjector_mock.o \
-    $(MOCK_DIR)/backend/utils/time/snapmgr_mock.o \
-    $(MOCK_DIR)/backend/access/heap/heapam_mock.o \
-    $(MOCK_DIR)/backend/access/common/heaptuple_mock.o \
-    $(MOCK_DIR)/backend/catalog/indexing_mock.o \
-    $(MOCK_DIR)/backend/access/common/scankey_mock.o \
-    $(MOCK_DIR)/backend/access/index/genam_mock.o
+    $(MOCK_DIR)/backend/utils/time/snapmgr_mock.o
