--- src/backend/fts/test/fts_test_helper.h
@@ -1,6 +1,0 @@
-#ifndef FTS_TEST_HELPER_H
-#define FTS_TEST_HELPER_H
-
-CdbComponentDatabases * InitTestCdb(int segCnt, bool has_mirrors, char default_mode);
-
-#endif
