+++ src/include/postmaster/ftsprobe.h
@@ -14,9 +14,63 @@
 #ifndef FTSPROBE_H
 #define FTSPROBE_H
 
-#include "cdb/ml_ipc.h" /* gettime_elapsed_ms */
+typedef struct
+{
+	int16 dbid;
+	bool isPrimaryAlive;
+	bool isMirrorAlive;
+	bool isInSync;
+	bool isSyncRepEnabled;
+	bool isRoleMirror;
+	bool retryRequested;
+} probe_result;
 
-void initPollFds(int size);
-void ftsConnect(fts_context *context);
-FtsProbeState nextFailedState(FtsProbeState state);
+/* States used by FTS main loop for probing segments. */
+typedef enum
+{
+	FTS_PROBE_SEGMENT,     /* send probe message */
+	FTS_SYNCREP_SEGMENT,   /* turn off syncrep due to mirror down */
+	FTS_PROMOTE_SEGMENT,   /* promote a mirror due to primary down */
+
+	/* wait before making another retry attempt */
+	FTS_PROBE_RETRY_WAIT,
+	FTS_SYNCREP_RETRY_WAIT,
+	FTS_PROMOTE_RETRY_WAIT,
+
+	FTS_PROBE_SUCCESS,     /* response to probe is ready for processing */
+	FTS_SYNCREP_SUCCESS,   /* syncrep was turned off by the primary */
+	FTS_PROMOTE_SUCCESS,   /* promotion was triggered on the mirror */
+
+	FTS_PROBE_FAILED,      /* the segment should be considered down */
+	FTS_SYNCREP_FAILED,    /* XXX how to handle this? */
+	FTS_PROMOTE_FAILED     /* double fault */
+} FtsProbeState;
+
+#define IsFtsProbeStateSuccess(state) (state == FTS_PROBE_SUCCESS || \
+		state == FTS_SYNCREP_SUCCESS || state == FTS_PROMOTE_SUCCESS)
+#define IsFtsProbeStateFailed(state) (state == FTS_PROBE_FAILED || \
+		state == FTS_SYNCREP_FAILED || state == FTS_PROMOTE_FAILED)
+typedef struct
+{
+	CdbComponentDatabaseInfo *primary_cdbinfo;
+	CdbComponentDatabaseInfo *mirror_cdbinfo;
+	probe_result result;
+	FtsProbeState state;
+	short poll_events;
+	short poll_revents;
+	int16 fd_index;               /* index into PollFds array */
+	pg_time_t startTime;          /* probe start timestamp */
+	pg_time_t retryStartTime;     /* time at which next retry attempt can start */
+	int16 probe_errno;            /* saved errno from the latest system call */
+	struct pg_conn *conn;         /* libpq connection object */
+	int retry_count;
+} per_segment_info;
+
+typedef struct
+{
+	int num_pairs; /* number of primary-mirror pairs FTS wants to probe */
+	per_segment_info *perSegInfos;
+} fts_context;
+
+extern bool FtsWalRepMessageSegments(CdbComponentDatabases *context);
 #endif
